requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
neo4j==5.15.0
httpx==0.25.2
PyGithub==2.1.1
apscheduler==3.10.4
python-multipart==0.0.6